+++ CHANGELOG.md
@@ -6,3 +6,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [0.0.2] - April 1, 2019
+
+### Changed
+
+- Transform and Build (replacing Construct and Predict).
+- Documentation updates.
+
+
+### Fixed
+
+- Hotfixes for documentation configuration.
+
+
+## [0.0.1] - March 31, 2019
+
+### Added
+
+- Initial release of alpha version.